LayoutTests/fast/forms/input-stepup-stepdown.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<script>
description('Check stepUp() and stepDown() bahevior for unsupported types.');

var input = document.createElement('input');
var invalidStateErr = '"InvalidStateError (DOM Exception 11): The object is in an invalid state."';

debug('Unsupported type');
input.type = 'text';
shouldThrowErrorName('input.step = "3"; input.min = ""; input.max = ""; input.value = "2"; input.stepDown()', 'InvalidStateError');
shouldThrowErrorName('input.stepDown(0)', 'InvalidStateError');
shouldThrowErrorName('input.stepUp()', 'InvalidStateError');
shouldThrowErrorName('input.stepUp(0)', 'InvalidStateError');
</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
